+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -655,5 +655,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-03T15:18:25.872Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-03T19:08:20.176Z")
     r"""The date and time the resource was last updated"""