+++ packages/orq-rc/RELEASES.md
@@ -2589,3 +2589,13 @@ Based on:
 - [python v3.12.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.35] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.35 - packages/orq-rc
+
+## 2025-09-03 19:09:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.609.0 (2.692.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.36 - packages/orq-rc